Adding a programatic approach example for requiring dsw. Closes #16
The problem with that is that the ./src/ files should be transpiled to work on previous versions of node!
So, if we use the `./src/main.js`, the command installed with -g (`npm install -g dsw`) will not work!
The command line program should run the `bin/index.js` while the require should load the `lib/dsw.js`.

For example, try and run this in your terminal:
```js
node
> const dsw = require(./lib/dsw.js);
> dsw.generate(./sandbox).then(result=>{ console.log(done, result); });
```
This should work.

But thanks for pointing this out. I just added `programatically.js` file so we can test it, in `sandbox`.